[#161] swagger 문서 섹션 분리
